Changed: #825 Remove all warnings when compiling Ryzom

+++ code/ryzom/client/src/interface_v3/group_html.cpp
@@ -3221,7 +3221,7 @@ void CGroupHTML::handle ()
 				else
 				{
 					/* Add our own request terminate handler. Nb: pass as param a UID, not the ptr */
-					HTNet_addAfter(requestTerminater, NULL, (void*)_GroupHtmlUID, HT_ALL, HT_FILTER_LAST);
+					HTNet_addAfter(requestTerminater, NULL, (void*)(size_t)_GroupHtmlUID, HT_ALL, HT_FILTER_LAST);
 
 					/* Set the timeout for long we are going to wait for a response */
 					HTHost_setEventTimeout(60000);
@@ -3367,7 +3367,7 @@ void CGroupHTML::handle ()
 				else
 				{
 					/* Add our own request terminate handler. Nb: pass as param a UID, not the ptr */
-					HTNet_addAfter(requestTerminater, NULL, (void*)_GroupHtmlUID, HT_ALL, HT_FILTER_LAST);
+					HTNet_addAfter(requestTerminater, NULL, (void*)(size_t)_GroupHtmlUID, HT_ALL, HT_FILTER_LAST);
 
 					/* Start the first request */
